eps-spring-batch/eps-spring-batch-ec-deployable/eps-batch/eps-batch-jobs/src/main/java/gov/hhs/cms/ff/fm/eps/ep/jobs/aop/ApplicationContextUtil.java
/**
 * 
 */
package gov.hhs.cms.ff.fm.eps.ep.jobs.aop;

import org.springframework.context.ApplicationContext;
import org.springframework.context.ApplicationContextAware;

/**
 * Utility class to get Spring Context
 * 
 * @author girish.padmanabhan
 */
public class ApplicationContextUtil implements ApplicationContextAware {
	private ApplicationContext context;
    
	public void setApplicationContext(ApplicationContext applicationContext) {
		context = applicationContext;
    }
	
    /**
     * @param requiredType
     * @return
     */
    public <T> T getBean(Class<T> requiredType) {
        return context.getBean(requiredType);
    }
    
    /**
     * @param name
     * @return
     */
    public Object getBean(String name) {
        return context.getBean(name);
    }

}
